+++ google/iam.go
@@ -466,3 +466,15 @@ func IamWithDeprecationMessage(message string) func(s *IamSettings) {
 func IamWithGAResourceDeprecation() func(s *IamSettings) {
 	return IamWithDeprecationMessage("This resource has been deprecated in the google (GA) provider, and will only be available in the google-beta provider in a future release.")
 }
+
+// Util to deref and print auditConfigs
+func debugPrintAuditConfigs(bs []*cloudresourcemanager.AuditConfig) string {
+	v, _ := json.MarshalIndent(bs, "", "\t")
+	return string(v)
+}
+
+// Util to deref and print bindings
+func debugPrintBindings(bs []*cloudresourcemanager.Binding) string {
+	v, _ := json.MarshalIndent(bs, "", "\t")
+	return string(v)
+}
